separate lattice config setting tabs logic
